add: freelst.c
