+++ includes/malloc.h
@@ -69,4 +69,10 @@ void	show_alloc_heap();
 void	show_free_list(t_region region);
 void	add_chunk_to_freelist(t_malloc_chunk *chunk, t_malloc_chunk **freelist);
 
+
+void	freelst_add_front(t_malloc_chunk **lst, t_malloc_chunk *new);
+void	freelst_insert(t_malloc_chunk *prev, t_malloc_chunk *new);
+void	freelst_replace(t_malloc_chunk *old, t_malloc_chunk *new);
+void	freelst_pop(t_malloc_chunk *lst);
+
 #endif
